Started work on a basic combat system
i don't even really want to talk about what i went through making this system.
General overview:
WeaponBase (or a derivative class) has multiple parameters which control how the weapon fires. this weapon is fired by calling on the fire() method. This spawns a bullet (what bullet is spawned is determined by the bullet provided as a prefab).
The bullet is part of a specific collision group, decided by which bullet prefab is used. The bullet will then fly until it hits a terrain object, or an object with the opposite collision group that it has. When it collides with something, it checks if that object has a HealthController. If it does, it deals an amount of damage. The bullet will delete itself no matter what.
I had a very very fun side adventure where i ran into a super strange issue. The position data of the bullets was being updated as they should, which was visible in the inspector. However, the bullet would in fact not visually move in the game view. Long story short, this was fixed by adding a ForceObjectToMove script to the parent object of the bullets. All this script does is add the identity quaternion to the node's rotation (so basically nothing), but by updating the rotation of the node all the children of the node update their visual position in the game view.
HealthController works as one would expect. You set a maxHealth value in the inspector, and can deal damage to it through the takeDamage() method.
God help my soul